fix interpolation

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -20,8 +20,8 @@ int interpolation_search(int *array, size_t size, int value)
 	high = size - 1;
 	while (size)
 	{
-		inter = (double)((array[high] - array[low]) * (value - array[low]));
-		pola = (double)((high - low) / inter);
+		inter = (array[high] - array[low]) * (value - array[low]);
+		pola = (double)(high - low) / inter;
 		pos = (size_t)(low + pola);
 		printf("Value checked array[%d]", (int)pos);
 		if (pos >= size)
